Extract IgnoreLogProcessor module
