+++ lib/log-processors.js
@@ -5,8 +5,8 @@ const LogProcessor = require('./LogProcessor')
 const JsonLogProcessor = require('./processors/JsonLogProcessor')
 const PrimitivesLogProcessor = require('./processors/PrimitivesLogProcessor')
 const SearchLogProcessor = require('./processors/SearchLogProcessor')
+const IgnoreLogProcessor = require('./processors/IgnoreLogProcessor')
 
-const { ignore } = builtInLogParsers
 const {
   appendLevel,
   appendMessage,
@@ -29,12 +29,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class IgnoreLogProcessor extends LogProcessor {
-  parse (input, context) {
-    return ignore(input, context)
   }
 }
